Refactor Artifact and add tests
Cleaned up Artifact class by removing unused variables and also removing
properties and methods that access the same data. This is so only
methods are used when getting file properties as there is the rare edge
case where a file property changes while swugenerator is running.
Caching file properties is probably ill advised because of edge cases.
Additionally, file lookup was moved to Artifact instantiation so
properties can be found without having to call a lookup method first.

In order to ensure the Artifact class behaves as expected and to
minimize the chances of regressions in the future, pytest unit tests
have been added for Artifact.

Signed-off-by: Colin McAllister <colinmca242@gmail.com>